+++ index.js
@@ -11,8 +11,6 @@ const rules = [
 module.exports = {
     extends: [
         'eslint-config-airbnb-base',
-        'plugin:vue/vue3-recommended',
-        '@vue/typescript/recommended',
         ...rules,
     ],
     env: {