+++ test/typescript.test.js
@@ -1,0 +1,14 @@
+const config = require('../typescript');
+
+const {
+    createEngine,
+    checkValidity,
+    checkLintingErrors,
+} = require('./test-utils');
+
+describe('base config', () => {
+    const engine = createEngine(config);
+
+    checkValidity(engine);
+    checkLintingErrors(engine);
+});