+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.1.0] - 2022-08-13
+### Added
+- Unit test for configuration validity
+
+### Fixed
+- Vue configurations inside the base configuration now properly reside in the Vue configuration only
+
 ## [3.0.0] - 2022-08-06
 ### Added
 - Support for Vue 3 via `@kellerkinder/eslint-config/vue`
@@ -88,7 +95,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Basic implementation based on AirBnB ESLint configuration
 
-[Unreleased]: https://github.com/kellerkinderDE/eslint-config/compare/3.0.0...HEAD
+[Unreleased]: https://github.com/kellerkinderDE/eslint-config/compare/3.1.0...HEAD
+[3.1.0]: https://github.com/kellerkinderDE/eslint-config/compare/3.0.0...3.1.0
 [3.0.0]: https://github.com/kellerkinderDE/eslint-config/compare/3.0.0-beta.1...3.0.0
 [3.0.0-beta.1]: https://github.com/kellerkinderDE/eslint-config/compare/2.2.0...3.0.0-beta.1
 [2.2.0]: https://github.com/kellerkinderDE/eslint-config/compare/2.1.0...2.2.0